Add community page (#176)
* Add community page

This commits adds the community page to the docs website.

* Improved text of Community page

This commit implements the requested changes for the new Community page and also adds some specific rules back to the matrix page.

Co-Authored-By: efb4f5ff-1298-471a-8973-3d47447115dc <73130443+efb4f5ff-1298-471a-8973-3d47447115dc@users.noreply.github.com>

* Move rules to child page of community page

This commit addresses the requested changes and moves the new rules page to a child page of the community page.

Co-Authored-By: efb4f5ff-1298-471a-8973-3d47447115dc <73130443+efb4f5ff-1298-471a-8973-3d47447115dc@users.noreply.github.com>
Co-Authored-By: Tim <67682506+ein-tim@users.noreply.github.com>

---------

Co-authored-by: efb4f5ff-1298-471a-8973-3d47447115dc <73130443+efb4f5ff-1298-471a-8973-3d47447115dc@users.noreply.github.com>